# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Petlove API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/usuarios/": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para buscar todos os usuários.",
        "responses": {
          "200": {
            "description": "Lista de usuários encontrada com sucesso.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "Nome do Usuário"
                  },
                  "email": {
                    "type": "string",
                    "example": "email@teste.com"
                  },
                  "senha": {
                    "type": "string",
                    "example": "hash_da_senha"
                  },
                  "foto": {
                    "type": "string",
                    "example": "http://localhost:8000/uploads/usuarios/foto.jpg"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00.000Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00.000Z"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para criar um novo usuário.",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "nome",
            "in": "formData",
            "type": "string",
            "description": "Nome do usuário."
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "description": "Email do usuário."
          },
          {
            "name": "senha",
            "in": "formData",
            "type": "string",
            "description": "Senha do usuário."
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "description": "Foto do usuário."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "foto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário criado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Usuario criado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para buscar um usuário pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "Nome do Usuário"
                },
                "email": {
                  "type": "string",
                  "example": "email@teste.com"
                },
                "senha": {
                  "type": "string",
                  "example": "hash_da_senha"
                },
                "foto": {
                  "type": "string",
                  "example": "http://localhost:8000/uploads/usuarios/foto.jpg"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-10-27T10:00:00.000Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023-10-27T10:00:00.000Z"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para editar um usuário existente.",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "nome",
            "in": "formData",
            "type": "string",
            "description": "Nome do usuário."
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "description": "Email do usuário."
          },
          {
            "name": "senha",
            "in": "formData",
            "type": "string",
            "description": "Senha do usuário."
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "description": "Foto do usuário."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "foto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário editado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Usuario atualizado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para deletar um usuário pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Usuario deletado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}